+++ lib/processor/http_test.go
@@ -69,6 +69,9 @@ func TestHTTPClientRetries(t *testing.T) {
 	if !HasFailed(msgs[0].Get(0)) {
 		t.Error("Failed message part not flagged")
 	}
+	if exp, act := "403", msgs[0].Get(0).Metadata().Get("http_processor_response_code"); exp != act {
+		t.Errorf("Wrong response code metadata: %v != %v", act, exp)
+	}
 
 	if exp, act := uint32(4), atomic.LoadUint32(&reqCount); exp != act {
 		t.Errorf("Wrong count of HTTP attempts: %v != %v", exp, act)
@@ -87,6 +90,8 @@ func TestHTTPClientBasic(t *testing.T) {
 			t.Errorf("Wrong payload value: %v != %v", act, exp)
 		}
 		i++
+		w.Header().Add("foobar", "baz")
+		w.WriteHeader(http.StatusCreated)
 		w.Write([]byte("foobar"))
 	}))
 	defer ts.Close()
@@ -106,6 +111,10 @@ func TestHTTPClientBasic(t *testing.T) {
 		t.Errorf("Wrong result count: %v != %v", actC, expC)
 	} else if exp, act := "foobar", string(message.GetAllBytes(msgs[0])[0]); act != exp {
 		t.Errorf("Wrong result: %v != %v", act, exp)
+	} else if exp, act := "201", msgs[0].Get(0).Metadata().Get("http_processor_response_code"); exp != act {
+		t.Errorf("Wrong response code metadata: %v != %v", act, exp)
+	} else if exp, act := "", msgs[0].Get(0).Metadata().Get("foobar"); exp != act {
+		t.Errorf("Wrong metadata value: %v != %v", act, exp)
 	}
 
 	msgs, res = h.ProcessMessage(message.New([][]byte{[]byte("bar")}))
@@ -115,6 +124,10 @@ func TestHTTPClientBasic(t *testing.T) {
 		t.Errorf("Wrong result count: %v != %v", actC, expC)
 	} else if exp, act := "foobar", string(message.GetAllBytes(msgs[0])[0]); act != exp {
 		t.Errorf("Wrong result: %v != %v", act, exp)
+	} else if exp, act := "201", msgs[0].Get(0).Metadata().Get("http_processor_response_code"); exp != act {
+		t.Errorf("Wrong response code metadata: %v != %v", act, exp)
+	} else if exp, act := "", msgs[0].Get(0).Metadata().Get("foobar"); exp != act {
+		t.Errorf("Wrong metadata value: %v != %v", act, exp)
 	}
 
 	// Check metadata persists.
@@ -129,6 +142,51 @@ func TestHTTPClientBasic(t *testing.T) {
 		t.Errorf("Wrong result: %v != %v", act, exp)
 	} else if exp, act := "bar", msgs[0].Get(0).Metadata().Get("foo"); exp != act {
 		t.Errorf("Metadata not preserved: %v != %v", act, exp)
+	} else if exp, act := "201", msgs[0].Get(0).Metadata().Get("http_processor_response_code"); exp != act {
+		t.Errorf("Wrong response code metadata: %v != %v", act, exp)
+	} else if exp, act := "", msgs[0].Get(0).Metadata().Get("foobar"); exp != act {
+		t.Errorf("Wrong metadata value: %v != %v", act, exp)
+	}
+}
+
+func TestHTTPClientBasicWithMetadata(t *testing.T) {
+	i := 0
+	expPayloads := []string{"foo", "bar", "baz"}
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		reqBytes, err := ioutil.ReadAll(r.Body)
+		if err != nil {
+			t.Fatal(err)
+		}
+		if exp, act := expPayloads[i], string(reqBytes); exp != act {
+			t.Errorf("Wrong payload value: %v != %v", act, exp)
+		}
+		i++
+		w.Header().Add("foobar", "baz")
+		w.WriteHeader(http.StatusCreated)
+		w.Write([]byte("foobar"))
+	}))
+	defer ts.Close()
+
+	conf := NewConfig()
+	conf.HTTP.Client.URL = ts.URL + "/testpost"
+	conf.HTTP.Client.CopyResponseHeaders = true
+
+	h, err := NewHTTP(conf, nil, log.Noop(), metrics.Noop())
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	msgs, res := h.ProcessMessage(message.New([][]byte{[]byte("foo")}))
+	if res != nil {
+		t.Error(res.Error())
+	} else if expC, actC := 1, msgs[0].Len(); actC != expC {
+		t.Errorf("Wrong result count: %v != %v", actC, expC)
+	} else if exp, act := "foobar", string(message.GetAllBytes(msgs[0])[0]); act != exp {
+		t.Errorf("Wrong result: %v != %v", act, exp)
+	} else if exp, act := "201", msgs[0].Get(0).Metadata().Get("http_processor_response_code"); exp != act {
+		t.Errorf("Wrong response code metadata: %v != %v", act, exp)
+	} else if exp, act := "baz", msgs[0].Get(0).Metadata().Get("foobar"); exp != act {
+		t.Errorf("Wrong metadata value: %v != %v", act, exp)
 	}
 }
 
@@ -138,6 +196,7 @@ func TestHTTPClientParallel(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		wg.Done()
 		wg.Wait()
+		w.WriteHeader(http.StatusCreated)
 		w.Write([]byte("foobar"))
 	}))
 	defer ts.Close()
@@ -168,6 +227,8 @@ func TestHTTPClientParallel(t *testing.T) {
 		t.Errorf("Wrong result: %v != %v", act, exp)
 	} else if exp, act := "bar", msgs[0].Get(0).Metadata().Get("foo"); exp != act {
 		t.Errorf("Metadata not preserved: %v != %v", act, exp)
+	} else if exp, act := "201", msgs[0].Get(0).Metadata().Get("http_processor_response_code"); exp != act {
+		t.Errorf("Wrong response code metadata: %v != %v", act, exp)
 	}
 }
 
@@ -218,12 +279,18 @@ func TestHTTPClientParallelError(t *testing.T) {
 	if !HasFailed(msgs[0].Get(2)) {
 		t.Error("Expected failed flag")
 	}
+	if exp, act := "403", msgs[0].Get(2).Metadata().Get("http_processor_response_code"); exp != act {
+		t.Errorf("Wrong response code metadata: %v != %v", act, exp)
+	}
 	for _, i := range []int{0, 1, 3, 4} {
 		if exp, act := "foobar", string(msgs[0].Get(i).Get()); act != exp {
 			t.Errorf("Wrong result: %v != %v", act, exp)
 		}
 		if HasFailed(msgs[0].Get(i)) {
 			t.Error("Did not expect failed flag")
+		}
+		if exp, act := "200", msgs[0].Get(i).Metadata().Get("http_processor_response_code"); exp != act {
+			t.Errorf("Wrong response code metadata: %v != %v", act, exp)
 		}
 	}
 }