+++ CHANGELOG.md
@@ -8,10 +8,14 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - New rate limit and websocket message fields added to `http_server` input.
+- The `http` processor now optionally copies headers from response into
+  resulting message metadata.
+- The `http` processor now sets a `http_processor_response_code` metadata value
+  into resulting messages.
 
 ### Changed
 
-- Go API: Removed experimental `Block` functions for the cache and rate limit
+- Go API: Removed experimental `Block` functions from the cache and rate limit
   packages.
 
 ## 2.9.0 - 2019-07-12